added a service folder and a class addformservices with the method for form verification and simplification of the method add
